RegistryAuthServer/requirements.txt
flask
pyjwt
cryptography
psycopg2
bcrypt
gunicorn
